+++ modules/maven/core/src/sap/commerce/toolset/maven/descriptor/MavenModuleDescriptor.kt
@@ -18,6 +18,7 @@
 
 package sap.commerce.toolset.maven.descriptor
 
+import com.intellij.openapi.project.Project
 import org.jetbrains.idea.maven.model.MavenConstants
 import sap.commerce.toolset.HybrisConstants
 import sap.commerce.toolset.project.descriptor.HybrisProjectDescriptor
@@ -33,7 +34,7 @@ class MavenModuleDescriptor(
 ) : ExternalModuleDescriptor(moduleRootDirectory, rootProjectDescriptor, moduleRootDirectory.name) {
 
     class Provider : ModuleDescriptorProvider {
-        override fun isApplicable(moduleRootDirectory: File): Boolean {
+        override fun isApplicable(project: Project?, moduleRootDirectory: File): Boolean {
             if (moduleRootDirectory.absolutePath.contains(HybrisConstants.PLATFORM_MODULE_PREFIX)) return false
 
             return File(moduleRootDirectory, MavenConstants.POM_XML).isFile()